requirements.txt
flask==3.1.3
werkzeug==3.1.6
pytest==8.3.5
pytest-flask==1.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
